backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0
PyJWT==2.9.0
mysql-connector-python==9.0.0
